fix: use timezone-aware timestamp() instead of timetuple() (#350)
fixes issue where convert.time ignores timezone in the datetime object.

current implementation:

times = [
     datetime(2020, 3, 14, 8, 0, tzinfo=timezone.utc),
     datetime(2020, 3, 14, 9, 0, tzinfo=timezone(timedelta(hours=1))),
     datetime(2020, 3, 14, 8, 0, tzinfo=timezone(timedelta(hours=1)))
]
for time in times:
	print(convert.time(time))

output:

1584153000
1584156600
1584153000

technically, as @atollena describes, the first two times should be
considered to be the same, whereas the third datetime instance is
expected be an hour behind the first two instances. but by using
timetuple(), the specified timezone is ignored. so, using the
timestamp() method instead of timetuple() ensures that the specified
timezone is respected:

new implementation (assuming the same list of times from before):

for time in times:
	print(convert.time(time))

output:

1584172800
1584172800
1584169200

here, the specified timezone data is respected, and as expected the
third datetime instance is 3600 seconds behind the first two instances.

since python2 support has now been dropped, this can safely be included
in the package.

Signed-off-by: Chinmay D. Pai <chinmaydpai@gmail.com>